requirement.txt
paho-mqtt
PyQt5